Added ViewTestCase.
