+++ st3/sublime_lib/testing.py
@@ -1,0 +1,19 @@
+import sublime
+
+from unittest import TestCase
+
+class ViewTestCase(TestCase):
+    def setUp(self):
+        self.view = sublime.active_window().new_file()
+
+    def tearDown(self):
+        if self.view:
+            self.view.set_scratch(True)
+            self.view.window().focus_view(self.view)
+            self.view.window().run_command("close_file")
+
+    def assertContents(self, text):
+        self.assertEqual(
+            self.view.substr(sublime.Region(0, self.view.size())),
+            text
+        )
